Make functions for all for operator buttons
Made a 'operatorProperties' function which will be triggered each time a
operator (+, -, *, divide) buttons are pressed. This function will save
the text on the screen before the operator to one of the variable(num1,
num2), save the current pressed operator to 'operator' variable and then
call 'equalTo' function(equalTo function is controlling another function
named 'operate' function which is doing all the mathematical stuff).

'operatorProperties' function is also calling 'display' function to
remove previous text and show result on screen. Currently
'operatorProperties' function has some bugs due to which results are not
correct!

We have also added one if block of code in the callback function of
'addEventListener' part!

Currently "=" button is not working.